+++ roles/pool/Cargo.toml
@@ -6,6 +6,10 @@ description = "SV2 pool role"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/stratum-mining/stratum"
 
+[lib]
+name = "pool_sv2"
+path = "src/lib/mod.rs"
+
 [dependencies]
 stratum-common = { path = "../../common" }
 async-channel = "1.5.1"
@@ -27,7 +31,6 @@ error_handling = { version = "0.1", path = "../../utils/error-handling" }
 nohash-hasher = "0.2.0"
 key-utils = { version = "^1.0.0", path = "../../utils/key-utils" }
 secp256k1 = { version = "0.27.0", default-features = false, features =["bitcoin_hashes","alloc","rand","rand-std"] }
-
 
 [dev-dependencies]
 hex = "0.4.3"
