+++ roles/pool/src/lib/error.rs
