+++ roles/pool/src/lib/mod.rs
@@ -1,2 +1,4 @@
 pub mod mining_pool;
 pub mod template_receiver;
+pub mod error;
+pub mod status;
