+++ roles/pool/src/lib/status.rs
@@ -1,6 +1,6 @@
 use roles_logic_sv2::parsers::Mining;
 
-use crate::error::PoolError;
+use super::error::PoolError;
 
 /// Each sending side of the status channel
 /// should be wrapped with this enum to allow
